fix: fix bfd static route tests (#12995)
Add fixes and improvements to bfd/test_bfd_static_route.py.
