+++ tests/bfd/conftest.py
@@ -1,6 +1,7 @@
 import pytest
 from bfd_base import BfdBase
 import logging
+
 from tests.platform_tests.link_flap.link_flap_utils import check_orch_cpu_utilization
 from tests.common.utilities import wait_until
 from tests.common.config_reload import config_reload
@@ -47,8 +48,8 @@ def bfd_cleanup_db(
     )
     for dut in duts:
         assert wait_until(
-            45, 2, 0, check_orch_cpu_utilization, dut, orch_cpu_threshold
-        ), "Orch CPU utilization {} > orch cpu threshold {} before starting the test".format(
+            120, 4, 0, check_orch_cpu_utilization, dut, orch_cpu_threshold
+        ), "Orch CPU utilization {} > orch cpu threshold {} after the test".format(
             dut.shell("show processes cpu | grep orchagent | awk '{print $9}'")[
                 "stdout"
             ],
@@ -87,6 +88,7 @@ def bfd_cleanup_db(
     request.config.dst_dut.shell(command)
 
     logger.info("Bringing up portchannels or respective members")
+    portchannels_on_dut = None
     if hasattr(request.config, "portchannels_on_dut"):
         portchannels_on_dut = request.config.portchannels_on_dut
         selected_interfaces = request.config.selected_portchannels
@@ -111,4 +113,32 @@ def bfd_cleanup_db(
             else request.config.dst_asic
         )
         for interface in selected_interfaces:
-            bfd_base_instance.interface_cleanup(dut, asic, interface)
+            ensure_interface_is_up(dut, asic, interface)
+
+
+def ensure_interface_is_up(dut, asic, interface):
+    int_oper_status = dut.show_interface(
+        command="status", include_internal_intfs=True, asic_index=asic.asic_index
+    )["ansible_facts"]["int_status"][interface]["oper_state"]
+    if int_oper_status == "down":
+        logger.info(
+            "Starting downed interface {} on {} asic{}".format(interface, dut, asic.asic_index)
+        )
+        exec_cmd = (
+            "sudo ip netns exec asic{} config interface -n asic{} startup {}".format(
+                asic.asic_index, asic.asic_index, interface
+            )
+        )
+
+        logger.info("Command: {}".format(exec_cmd))
+        dut.shell(exec_cmd)
+        assert wait_until(
+            180,
+            10,
+            0,
+            lambda: dut.show_interface(
+                command="status",
+                include_internal_intfs=True,
+                asic_index=asic.asic_index,
+            )["ansible_facts"]["int_status"][interface]["oper_state"] == "up",
+        )
